@@ -2,7 +2,20 @@
 
 require_relative 'boot'
 
-require 'rails/all'
+require 'rails'
+# Pick the frameworks you want:
+require 'active_model/railtie'
+require 'active_job/railtie'
+require 'active_record/railtie'
+# require 'active_storage/engine'
+require 'action_controller/railtie'
+require 'action_mailer/railtie'
+# require 'action_mailbox/engine'
+# require 'action_text/engine'
+require 'action_view/railtie'
+# require 'action_cable/engine'
+# require "sprockets/railtie"
+require 'rails/test_unit/railtie'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -18,18 +31,16 @@ module Martigua2
                        routing_specs: false,
                        controller_specs: false,
                        request_specs: false
-      g.fixture_replacement :factory_bot, dir: "spec/factories"
+      g.fixture_replacement :factory_bot, dir: 'spec/factories'
     end
 
     config.time_zone = 'Paris'
     config.active_record.default_timezone = :local
 
-    config.active_job.queue_adapter = :delayed_job
-
     config.middleware.insert_before 0, Rack::Cors do
       allow do
         origins 'd1zljy12d9ls0t.cloudfront.net'
-        resource '*', :headers => :any, :methods => [:get, :head, :options]
+        resource '*', headers: :any, methods: %i[get head options]
       end
     end
 
@@ -40,7 +51,7 @@ module Martigua2
       end
     end
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.2
+    config.load_defaults 6.0
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
